requirements.txt
python-cmr==0.11.0
matplotlib==3.9.1
tqdm
